Extract Tx into is own module
