+++ wallet-api/wallet-api.cabal
@@ -63,6 +63,7 @@ library
         Ledger.Slot
         Ledger.Scripts
         Ledger.Types
+        Ledger.Tx
         Ledger.Validation
         Ledger.Index
         Ledger.Interval
